Cleanup switching, add subtitle switching
Aiming for a UI similar to mplayer, I would like to have '#' to
switch the audio track. However, from the Navigation Events, for
a US keyboard, I get Shift+3. For a German keyboard layout,
I get 'numbersign'.

All in all, I should either decide to ignore
Keyboard layouts alltogether (Shift+3) or I should try to get from
the key-release events to a real, written character.
The former would allow a nice keyboard usage (If you ever used
mplayer with a German keyboard, you notice the character mapping
was chosen with the US keyboard layout in mind), however it is
condemned to confuse users. On the other hand, if there isn't another
GStreamer interface which provides this information directly, the
latter is probably quite hard to do, especially when considering that
I would like to keep this player sink-agnostic. (Yea, that dilusion
will probably die at some point on the way, but for now, I would like
to keep it up ;) )